refactor(npm): move InNpmPackageChecker code to deno_resolver
