+++ resolvers/deno/sync.rs
@@ -6,6 +6,8 @@ pub use inner::*;
 mod inner {
   #![allow(clippy::disallowed_types)]
 
+  pub use core::marker::Send as MaybeSend;
+  pub use core::marker::Sync as MaybeSync;
   pub use std::sync::Arc as MaybeArc;
 
   pub use dashmap::DashMap as MaybeDashMap;
@@ -20,6 +22,11 @@ mod inner {
   use std::hash::Hash;
   use std::hash::RandomState;
   pub use std::rc::Rc as MaybeArc;
+
+  pub trait MaybeSync {}
+  impl<T> MaybeSync for T where T: ?Sized {}
+  pub trait MaybeSend {}
+  impl<T> MaybeSend for T where T: ?Sized {}
 
   // Wrapper struct that exposes a subset of `DashMap` API.
   #[derive(Debug)]
